Cast numbers to avoid some compiler warnings.
git-svn-id: https://svn.dealii.org/trunk@13079 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/source/multigrid/mg_dof_accessor.cc
@@ -33,8 +33,8 @@ template <int structdim, int dim>
 MGDoFAccessor<structdim, dim>::MGDoFAccessor ()
 		:
 		BaseClass (0,
-			   deal_II_numbers::invalid_unsigned_int,
-			   deal_II_numbers::invalid_unsigned_int,
+			   static_cast<int>(deal_II_numbers::invalid_unsigned_int),
+			   static_cast<int>(deal_II_numbers::invalid_unsigned_int),
 			   0),
 		mg_dof_handler(0)
 {
